% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_cube.R
\name{init_cube}
\alias{init_cube}
\title{Initialize a pocket cube (cube of size 2x2x2)}
\usage{
init_cube()
}
\value{
A string representing the solved cube
}
\description{
A string is used to represent the current state of the cube. It is not recommended for users to manually modify the string expect for special situations.
}
\examples{
init_cube()

}
\seealso{
\code{\link{op}}
}
