% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cube.R
\name{is_cube}
\alias{is_cube}
\title{Determining if this string can represent a pocket cube}
\usage{
is_cube(x)
}
\arguments{
\item{x}{A string}
}
\value{
Whether there is a state of the pocket cube that corresponds to \code{x}.
}
\description{
Although it is not recommended for users to manually edit strings to represent the state of a cube, we provide a function to check if a string can represent a specific state.
}
\examples{
is_cube(init_cube())

}
\seealso{
\code{\link{cube_solve}}
}
