% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NISS.R
\name{NISS}
\alias{NISS}
\title{Find the inverse scramble}
\usage{
NISS(scramble)
}
\arguments{
\item{scramble}{A sequence for scrambling. See Details for more information}
}
\value{
The inverse scramble
}
\description{
Reverse the order of the scramble and invert each of the individual turns. For instance, \code{NISS("RUF")="F'U'R'"}.
}
\details{
The most popular and standard rotation notation is used. Turns for a pocket Cube are represented by the following moves: R, R2, R', U, U2, U', F, F2, F'.
These moves correspond to the following actions:

R: clockwise rotation of the right face

R2: double (180 degrees) clockwise rotation of the right face

R': counterclockwise rotation of the right face

U: clockwise rotation of the upper face

U2: double (180 degrees) clockwise rotation of the upper face

U': counterclockwise rotation of the upper face

F: clockwise rotation of the front face

F2: double (180 degrees) clockwise rotation of the front face

F': counterclockwise rotation of the front face

When multiple moves are performed in succession, they are executed from left to right.
}
\examples{
x=randcube()
#Find the solution
sol=cube_solve(x)
#Find the scramble sequence
NISS(sol)

}
\seealso{
\code{\link{cube_solve}}  \code{\link{op}}
}
