% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcube.R
\name{plotcube}
\alias{plotcube}
\title{Plot the cube in 2D or 3D}
\usage{
plotcube(x, type = "2D")
}
\arguments{
\item{x}{A string representing the current state of the pocket cube. For the solved state, the string returned by \code{\link{init_cube}} can be used.}

\item{type}{Whether the plot is 2D or 3D. Default is \code{"2D"}.}
}
\description{
This function refers to rcube:::plot.cube and \code{rcube::plot3dFlat} to plot the cube.
}
\examples{
x=init_cube()
plotcube(x)
y=randcube()
plotcube(y,type="3D")

}
\seealso{
\code{\link{cube_solve}}  \code{\link{randcube}}
}
