% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randcube.R
\name{randcube}
\alias{randcube}
\title{Randomly generate a scrambled cube}
\usage{
randcube()
}
\value{
A string representing the current state of the cube
}
\description{
Randomly generate a scrambled cube
}
\examples{
x=randcube()
#Find the corresponding scramble sequence
NISS(cube_solve(x))
plotcube(x,type="3D")

}
\seealso{
\code{\link{cube_solve}}  \code{\link{plotcube}} \code{\link{NISS}}
}
